+++ servant-uverb-swagger/servant-uverb-swagger.cabal
@@ -1,4 +1,4 @@
-cabal-version:       2.0
+cabal-version:       2.2
 name:                servant-uverb-swagger
 version:             0.0.1
 synopsis:            See README for more info
@@ -19,14 +19,8 @@ source-repository head
   type:                git
   location:            https://github.com/wireapp/servant-uverb.git
 
-library
-  hs-source-dirs:      src
-  exposed-modules:     Servant.UVerb.Swagger
-
-
-  build-depends:       base ^>= 4.12.0.0
-
-
+common standard-things 
+  build-depends:        base ^>= 4.12.0.0
   ghc-options:         -Wall
                        -Wincomplete-uni-patterns
                        -Wincomplete-record-updates
@@ -36,13 +30,12 @@ library
                        -fhide-source-paths
                        -Wmissing-export-lists
                        -Wpartial-fields
-
   default-language:    Haskell2010
   default-extensions:  ConstraintKinds
-                       AllowAmbiguousTypes
                        DataKinds
                        DeriveFunctor
                        DeriveGeneric
+                       DerivingStrategies
                        DerivingVia
                        DuplicateRecordFields
                        FlexibleContexts
@@ -62,49 +55,18 @@ library
                        UndecidableInstances
                        ViewPatterns
 
+library
+  import:              standard-things
+  hs-source-dirs:      src
+  exposed-modules:     Servant.Swagger.UVerb
+
 test-suite servant-uverb-test
+  import:              standard-things
   type:                exitcode-stdio-1.0
   hs-source-dirs:      test
   main-is:             Spec.hs
-
-  build-depends:       base ^>= 4.12.0.0
-                     , servant-uverb
-
-
+  build-depends:       servant-uverb
   ghc-options:         -Wall
                        -threaded
                        -rtsopts
                        -with-rtsopts=-N
-                       -Wincomplete-uni-patterns
-                       -Wincomplete-record-updates
-                       -Wcompat
-                       -Widentities
-                       -Wredundant-constraints
-                       -fhide-source-paths
-                       -Wmissing-export-lists
-                       -Wpartial-fields
-
-  default-language:    Haskell2010
-  default-extensions:  ConstraintKinds
-                       AllowAmbiguousTypes
-                       DataKinds
-                       DeriveFunctor
-                       DeriveGeneric
-                       DerivingVia
-                       DuplicateRecordFields
-                       FlexibleContexts
-                       GeneralizedNewtypeDeriving
-                       InstanceSigs
-                       KindSignatures
-                       LambdaCase
-                       OverloadedStrings
-                       PolyKinds
-                       RankNTypes
-                       RecordWildCards
-                       ScopedTypeVariables
-                       StandaloneDeriving
-                       TupleSections
-                       TypeApplications
-                       TypeOperators
-                       UndecidableInstances
-                       ViewPatterns
