+++ servant-uverb-swagger/src/Servant/Swagger/UVerb.hs
@@ -1,4 +1,4 @@
-module Servant.UVerb.Client
+module Servant.Swagger.UVerb
        ( someFunc
        ) where
